model.TemplateReference -> TemplateRef
